moved method to utils
